# fix(#616): Implement correct Steel Joker as deck composition multiplier