--- a/core/benches/effect_processor_benchmark.rs
+++ b/core/benches/effect_processor_benchmark.rs
@@ -352,6 +352,7 @@ fn create_test_game_context() -> GameContext<'static> {
         stage,
         hands_played: 0,
         discards_used: 0,
+        hands_remaining: 4.0, // Standard hands remaining for testing
         jokers,
         hand,
         discarded,