--- a/core/benches/effect_processor_benchmark_simple.rs
+++ b/core/benches/effect_processor_benchmark_simple.rs
@@ -145,6 +145,7 @@ fn create_test_game_context() -> balatro_rs::joker::GameContext<'static> {
         stage,
         hands_played: 0,
         discards_used: 0,
+        hands_remaining: 4.0, // Standard hands remaining for testing
         jokers,
         hand,
         discarded,