--- a/core/benches/joker_effect_processor_trait_optimization.rs
+++ b/core/benches/joker_effect_processor_trait_optimization.rs
@@ -48,6 +48,7 @@ impl TestGameData {
             stage: &self.stage,
             hands_played: 0,
             discards_used: 0,
+            hands_remaining: 4.0, // Standard hands remaining for testing
             jokers: &[],
             hand: &self.hand,
             discarded: &[],