--- a/core/benches/trait_benchmark.rs
+++ b/core/benches/trait_benchmark.rs
@@ -378,6 +378,7 @@ impl TestGameData {
             stage: &self.stage,
             hands_played: 0,
             discards_used: 0,
+            hands_remaining: 4.0, // Standard hands remaining for testing
             jokers: &[],
             hand: &self.hand,
             discarded: &[],