--- a/core/src/game/mod.rs
+++ b/core/src/game/mod.rs
@@ -5,7 +5,6 @@ use crate::boss_blinds::BossBlindState;
 use crate::bounded_action_history::BoundedActionHistory;
 use crate::card::Card;
 use crate::config::Config;
-use crate::consumables::ConsumableError;
 use crate::consumables::ConsumableId;
 use crate::deck::Deck;
 use crate::error::GameError;
@@ -24,7 +23,6 @@ use crate::stage::{Blind, End, Stage};
 use crate::state_version::StateVersion;
 use crate::target_context::TargetContext;
 use crate::vouchers::{VoucherCollection, VoucherId};
-use strum::IntoEnumIterator;
 
 // Re-export GameState for external use with qualified name to avoid Python bindings conflict
 pub use crate::vouchers::GameState as VoucherGameState;
@@ -212,16 +210,13 @@ pub struct Game {
     // Card enhancement tracking for this game run
     /// Count of Stone cards currently in deck (cached for performance)
     pub stone_cards_in_deck: usize,
-
+    
     /// Count of Steel cards currently in deck (cached for performance)
     pub steel_cards_in_deck: usize,
 
     // Extended state for consumables, vouchers, and boss blinds
     /// Consumable cards currently in the player's hand
     pub consumables_in_hand: Vec<ConsumableId>,
-
-    /// Consumable card slots for managing player consumable inventory
-    pub consumable_slots: crate::consumables::ConsumableSlots,
 
     /// Collection of owned vouchers with purchase tracking
     pub vouchers: VoucherCollection,
@@ -254,8 +249,6 @@ pub struct Game {
     /// Random number generator for secure game randomness
     #[cfg_attr(feature = "serde", serde(skip, default = "default_game_rng"))]
     pub rng: crate::rng::GameRng,
-    /// Hand levels for dynamic scoring - tracks current level for each hand type
-    pub hand_levels: HashMap<HandRank, usize>,
 
     /// Memory monitor for tracking and controlling memory usage
     #[cfg_attr(feature = "serde", serde(skip))]
@@ -354,14 +347,13 @@ impl Game {
             mult: config.base_mult as f64,
             score: config.base_score as f64,
             hand_type_counts: HashMap::new(),
-
+            
             // Initialize enhancement tracking (will be calculated after deck is set up)
             stone_cards_in_deck: 0,
             steel_cards_in_deck: 0,
 
             // Initialize extended state fields
             consumables_in_hand: Vec::new(),
-            consumable_slots: crate::consumables::ConsumableSlots::new(),
             vouchers: VoucherCollection::new(),
             boss_blind_state: BossBlindState::new(),
 
@@ -382,8 +374,6 @@ impl Game {
             rng: crate::rng::GameRng::secure(),
 
             // Initialize memory monitor with default configuration
-            // Initialize hand levels - all hands start at level 1
-            hand_levels: HandRank::iter().map(|hand| (hand, 1)).collect(),
             memory_monitor: MemoryMonitor::default(),
 
             // Initialize skip tags system
@@ -405,7 +395,7 @@ impl Game {
             .count()
     }
 
-    /// Count Steel cards in the current deck
+    /// Count Steel cards in the current deck  
     /// Following clean code principle: functions should do one thing
     fn count_steel_cards(&self) -> usize {
         self.deck
@@ -429,7 +419,7 @@ impl Game {
         self.deck.extend(cards);
     }
 
-    /// Get deck size for testing purposes
+    /// Get deck size for testing purposes  
     /// Following clean code: provide necessary test access without exposing internals
     #[cfg(test)]
     pub fn deck_size_for_testing(&self) -> usize {
@@ -439,41 +429,11 @@ impl Game {
     pub fn start(&mut self) {
         // for now just move state to small blind
         self.stage = Stage::PreBlind();
-
+        
         // Refresh enhancement counts after deck is set up
         self.refresh_enhancement_counts();
-
+        
         self.deal();
-    }
-
-    /// Get dynamic level information for a hand type based on current game state
-    pub fn get_hand_level(&self, hand_rank: HandRank) -> crate::rank::Level {
-        let current_level = self.hand_levels.get(&hand_rank).copied().unwrap_or(1);
-        let base_level = hand_rank.base_level();
-
-        // Calculate dynamic values based on current level
-        // Each level increases chips by base amount and mult by 1
-        let level_multiplier = current_level as f64;
-        crate::rank::Level {
-            level: current_level,
-            chips: (base_level.chips as f64 * level_multiplier) as usize,
-            mult: base_level.mult + (current_level - 1), // Add 1 mult per level above 1
-        }
-    }
-
-    /// Level up a specific hand type (used by planet cards)
-    pub fn level_up_hand(&mut self, hand_rank: HandRank) -> Result<(), ConsumableError> {
-        let current_level = self.hand_levels.get(&hand_rank).copied().unwrap_or(1);
-
-        // Cap level at reasonable maximum (e.g., 10)
-        if current_level >= 10 {
-            return Err(ConsumableError::InvalidGameState(format!(
-                "Hand {hand_rank:?} is already at maximum level"
-            )));
-        }
-
-        self.hand_levels.insert(hand_rank, current_level + 1);
-        Ok(())
     }
 
     /// Start a new blind and trigger joker lifecycle events
@@ -680,8 +640,8 @@ impl Game {
 
     pub fn calc_score(&mut self, hand: MadeHand) -> f64 {
         // compute chips and mult from hand level
-        self.chips += self.get_hand_level(hand.rank).chips as f64;
-        self.mult += self.get_hand_level(hand.rank).mult as f64;
+        self.chips += hand.rank.level().chips as f64;
+        self.mult += hand.rank.level().mult as f64;
 
         // add chips for each played card
         let card_chips: f64 = hand.hand.cards().iter().map(|c| c.chips() as f64).sum();
@@ -749,8 +709,8 @@ impl Game {
             joker_state_manager: &self.joker_state_manager,
             hand_type_counts: &self.hand_type_counts,
             cards_in_deck: self.deck.len(),
-            stone_cards_in_deck: self.stone_cards_in_deck,
-            steel_cards_in_deck: self.steel_cards_in_deck,
+            stone_cards_in_deck: 0, // TODO: Track stone cards when implemented
+            steel_cards_in_deck: 0, // TODO: Track steel cards when implemented
             rng: &self.rng,
         };
 
@@ -890,8 +850,8 @@ impl Game {
         let _initial_mult = self.mult;
 
         // Calculate base values from hand level
-        let base_chips = self.get_hand_level(hand.rank).chips as f64;
-        let base_mult = self.get_hand_level(hand.rank).mult as f64;
+        let base_chips = hand.rank.level().chips as f64;
+        let base_mult = hand.rank.level().mult as f64;
         self.chips += base_chips;
         self.mult += base_mult;
 
@@ -1673,9 +1633,7 @@ impl Game {
                             .unwrap_or(ConsumableId::Mercury)
                     }
                     crate::shop::ConsumableType::Spectral => {
-                        // Use All pool for regular shop purchases (includes Soul and Black Hole)
-                        // Note: Ghost Deck restrictions would be implemented at the shop level
-                        let spectral_cards = crate::consumables::SpectralPool::All.get_cards();
+                        let spectral_cards = ConsumableId::spectral_cards();
                         self.rng
                             .choose(&spectral_cards)
                             .copied()
@@ -1949,12 +1907,6 @@ impl Game {
                 // TODO: Implement multi-select deactivation
                 Err(GameError::InvalidAction)
             }
-
-            // Planet card actions
-            Action::UsePlanetCard {
-                planet_card_id,
-                hand_rank_id,
-            } => self.use_planet_card(planet_card_id, hand_rank_id),
             // Consumable actions - infrastructure ready for implementation
             Action::BuyConsumable {
                 consumable_id: _,
@@ -2445,14 +2397,13 @@ impl Game {
             mult: saveable_state.mult,
             score: saveable_state.score,
             hand_type_counts: saveable_state.hand_type_counts,
-
+            
             // Enhancement tracking (will be calculated after loading)
             stone_cards_in_deck: 0,
             steel_cards_in_deck: 0,
-
+            
             // Extended state fields
             consumables_in_hand: saveable_state.consumables_in_hand,
-            consumable_slots: crate::consumables::ConsumableSlots::new(),
             vouchers: saveable_state.vouchers,
             boss_blind_state: saveable_state.boss_blind_state,
             pack_inventory: saveable_state.pack_inventory,
@@ -2466,8 +2417,6 @@ impl Game {
             // Initialize secure RNG (not serialized)
             rng: crate::rng::GameRng::secure(),
             // Initialize memory monitor (not serialized)
-            // Initialize hand levels - all hands start at level 1
-            hand_levels: HandRank::iter().map(|hand| (hand, 1)).collect(),
             memory_monitor: MemoryMonitor::default(),
             // Initialize skip tags system (not serialized)
             available_skip_tags: Vec::new(),
@@ -2508,78 +2457,14 @@ impl Game {
             joker_state_manager: &self.joker_state_manager,
             hand_type_counts: &self.hand_type_counts,
             cards_in_deck: self.deck.len(),
-            stone_cards_in_deck: self.stone_cards_in_deck,
-            steel_cards_in_deck: self.steel_cards_in_deck,
+            stone_cards_in_deck: 0, // TODO: Track stone cards when implemented
+            steel_cards_in_deck: 0, // TODO: Track steel cards when implemented
             rng: &self.rng,
         };
 
         // Process scaling events for any scaling jokers
         for joker in &self.jokers {
             joker.process_scaling_event(&mut context, &event);
-        }
-    }
-}
-
-// Consumable functionality for Game
-impl Game {
-    /// Use a consumable from the players consumable slots
-    pub fn use_consumable(&mut self, _consumable_slot: usize) -> Result<(), GameError> {
-        // For now, return error until consumable_slots is implemented
-        // Get the consumable from the slot
-        if let Some(_consumable) = self.consumable_slots.get_consumable(_consumable_slot) {
-            // Use the consumable - for now just remove it and return success
-            // TODO: Implement actual consumable targeting and effects
-            if let Ok(_removed_consumable) =
-                self.consumable_slots.remove_consumable(_consumable_slot)
-            {
-                // Consumable was used successfully
-                Ok(())
-            } else {
-                Err(GameError::InvalidAction)
-            }
-        } else {
-            Err(GameError::InvalidAction)
-        }
-    }
-
-    /// Use a planet card to level up a hand type
-    pub fn use_planet_card(
-        &mut self,
-        planet_card_id: u32,
-        hand_rank_id: u32,
-    ) -> Result<(), GameError> {
-        use crate::consumables::planet::create_planet_card;
-        use crate::rank::HandRank;
-        use strum::IntoEnumIterator;
-
-        // Convert IDs back to proper types
-        let planet_card = crate::consumables::ConsumableId::iter()
-            .nth(planet_card_id as usize)
-            .ok_or(GameError::InvalidAction)?;
-        let hand_rank = HandRank::iter()
-            .nth(hand_rank_id as usize)
-            .ok_or(GameError::InvalidAction)?;
-
-        // Validate that this is a planet card
-        if !matches!(
-            planet_card.consumable_type(),
-            crate::consumables::ConsumableType::Planet
-        ) {
-            return Err(GameError::InvalidAction);
-        }
-
-        // Create the planet card instance and use it
-        if let Some(card) = create_planet_card(planet_card) {
-            let target = crate::consumables::Target::HandType(hand_rank);
-            if card.can_use(self, &target) {
-                card.use_effect(self, target)
-                    .map_err(|_| GameError::InvalidAction)?;
-                Ok(())
-            } else {
-                Err(GameError::InvalidAction)
-            }
-        } else {
-            Err(GameError::InvalidAction)
         }
     }
 }
@@ -4220,7 +4105,7 @@ mod tests {
 #[cfg(test)]
 mod stone_steel_tracking_tests {
     use super::*;
-    use crate::card::{Card, Edition, Enhancement, Suit, Value};
+    use crate::card::{Card, Enhancement, Suit, Value, Edition};
     use crate::config::Config;
 
     /// Test helper to create a card with specific enhancement
@@ -4239,50 +4124,50 @@ mod stone_steel_tracking_tests {
     fn should_count_zero_stone_cards_in_empty_deck() {
         let mut game = Game::new(Config::default());
         game.refresh_enhancement_counts();
-
+        
         assert_eq!(game.stone_cards_in_deck, 0);
     }
 
     #[test]
     fn should_count_one_stone_card_in_deck() {
         let mut game = Game::new(Config::default());
-
+        
         let stone_card = create_card_with_enhancement(Some(Enhancement::Stone));
         game.add_cards_to_deck_for_testing(vec![stone_card]);
-
+        
         game.refresh_enhancement_counts();
-
+        
         assert_eq!(game.stone_cards_in_deck, 1);
     }
 
     #[test]
     fn should_count_multiple_stone_cards_in_deck() {
         let mut game = Game::new(Config::default());
-
+        
         let stone_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Stone)),
         ];
         game.add_cards_to_deck_for_testing(stone_cards);
-
+        
         game.refresh_enhancement_counts();
-
+        
         assert_eq!(game.stone_cards_in_deck, 3);
     }
 
-    #[test]
+    #[test] 
     fn should_count_zero_steel_cards_in_empty_deck() {
         let mut game = Game::new(Config::default());
         game.refresh_enhancement_counts();
-
+        
         assert_eq!(game.steel_cards_in_deck, 0);
     }
 
     #[test]
     fn should_count_multiple_steel_cards_in_deck() {
         let mut game = Game::new(Config::default());
-
+        
         let steel_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Steel)),
             create_card_with_enhancement(Some(Enhancement::Steel)),
@@ -4290,16 +4175,16 @@ mod stone_steel_tracking_tests {
             create_card_with_enhancement(Some(Enhancement::Steel)),
         ];
         game.add_cards_to_deck_for_testing(steel_cards);
-
+        
         game.refresh_enhancement_counts();
-
+        
         assert_eq!(game.steel_cards_in_deck, 4);
     }
 
     #[test]
     fn should_count_both_stone_and_steel_cards_independently() {
         let mut game = Game::new(Config::default());
-
+        
         let mixed_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Steel)),
@@ -4309,9 +4194,9 @@ mod stone_steel_tracking_tests {
             create_card_with_enhancement(Some(Enhancement::Bonus)), // Should be ignored
         ];
         game.add_cards_to_deck_for_testing(mixed_cards);
-
+        
         game.refresh_enhancement_counts();
-
+        
         assert_eq!(game.stone_cards_in_deck, 2);
         assert_eq!(game.steel_cards_in_deck, 3);
     }
@@ -4319,20 +4204,20 @@ mod stone_steel_tracking_tests {
     #[test]
     fn should_refresh_enhancement_counts_on_game_start() {
         let mut game = Game::new(Config::default());
-
+        
         let enhanced_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Steel)),
         ];
         game.add_cards_to_deck_for_testing(enhanced_cards);
-
+        
         // Counts should be 0 before start (not yet calculated)
         assert_eq!(game.stone_cards_in_deck, 0);
         assert_eq!(game.steel_cards_in_deck, 0);
-
+        
         // Start the game (this should refresh counts)
         game.start();
-
+        
         // Counts should now be accurate
         assert_eq!(game.stone_cards_in_deck, 1);
         assert_eq!(game.steel_cards_in_deck, 1);
@@ -4341,9 +4226,9 @@ mod stone_steel_tracking_tests {
     #[test]
     fn should_handle_deck_size_calculation_correctly() {
         let mut game = Game::new(Config::default());
-
+        
         let initial_deck_size = game.deck_size_for_testing();
-
+        
         let cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Stone)),
@@ -4353,7 +4238,7 @@ mod stone_steel_tracking_tests {
         ];
         game.add_cards_to_deck_for_testing(cards);
         game.refresh_enhancement_counts();
-
+        
         // Total deck size should be initial size + 5 added cards
         assert_eq!(game.deck_size_for_testing(), initial_deck_size + 5);
         // Enhancement counts should be accurate (only counting the added enhanced cards)