--- a/core/src/game/mod.rs
+++ b/core/src/game/mod.rs
@@ -210,7 +210,7 @@ pub struct Game {
     // Card enhancement tracking for this game run
     /// Count of Stone cards currently in deck (cached for performance)
     pub stone_cards_in_deck: usize,
-    
+
     /// Count of Steel cards currently in deck (cached for performance)
     pub steel_cards_in_deck: usize,
 
@@ -347,7 +347,7 @@ impl Game {
             mult: config.base_mult as f64,
             score: config.base_score as f64,
             hand_type_counts: HashMap::new(),
-            
+
             // Initialize enhancement tracking (will be calculated after deck is set up)
             stone_cards_in_deck: 0,
             steel_cards_in_deck: 0,
@@ -429,10 +429,10 @@ impl Game {
     pub fn start(&mut self) {
         // for now just move state to small blind
         self.stage = Stage::PreBlind();
-        
+
         // Refresh enhancement counts after deck is set up
         self.refresh_enhancement_counts();
-        
+
         self.deal();
     }
 
@@ -2397,11 +2397,11 @@ impl Game {
             mult: saveable_state.mult,
             score: saveable_state.score,
             hand_type_counts: saveable_state.hand_type_counts,
-            
+
             // Enhancement tracking (will be calculated after loading)
             stone_cards_in_deck: 0,
             steel_cards_in_deck: 0,
-            
+
             // Extended state fields
             consumables_in_hand: saveable_state.consumables_in_hand,
             vouchers: saveable_state.vouchers,
@@ -4105,7 +4105,7 @@ mod tests {
 #[cfg(test)]
 mod stone_steel_tracking_tests {
     use super::*;
-    use crate::card::{Card, Enhancement, Suit, Value, Edition};
+    use crate::card::{Card, Edition, Enhancement, Suit, Value};
     use crate::config::Config;
 
     /// Test helper to create a card with specific enhancement
@@ -4124,50 +4124,50 @@ mod stone_steel_tracking_tests {
     fn should_count_zero_stone_cards_in_empty_deck() {
         let mut game = Game::new(Config::default());
         game.refresh_enhancement_counts();
-        
+
         assert_eq!(game.stone_cards_in_deck, 0);
     }
 
     #[test]
     fn should_count_one_stone_card_in_deck() {
         let mut game = Game::new(Config::default());
-        
+
         let stone_card = create_card_with_enhancement(Some(Enhancement::Stone));
         game.add_cards_to_deck_for_testing(vec![stone_card]);
-        
+
         game.refresh_enhancement_counts();
-        
+
         assert_eq!(game.stone_cards_in_deck, 1);
     }
 
     #[test]
     fn should_count_multiple_stone_cards_in_deck() {
         let mut game = Game::new(Config::default());
-        
+
         let stone_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Stone)),
         ];
         game.add_cards_to_deck_for_testing(stone_cards);
-        
+
         game.refresh_enhancement_counts();
-        
+
         assert_eq!(game.stone_cards_in_deck, 3);
     }
 
-    #[test] 
+    #[test]
     fn should_count_zero_steel_cards_in_empty_deck() {
         let mut game = Game::new(Config::default());
         game.refresh_enhancement_counts();
-        
+
         assert_eq!(game.steel_cards_in_deck, 0);
     }
 
     #[test]
     fn should_count_multiple_steel_cards_in_deck() {
         let mut game = Game::new(Config::default());
-        
+
         let steel_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Steel)),
             create_card_with_enhancement(Some(Enhancement::Steel)),
@@ -4175,16 +4175,16 @@ mod stone_steel_tracking_tests {
             create_card_with_enhancement(Some(Enhancement::Steel)),
         ];
         game.add_cards_to_deck_for_testing(steel_cards);
-        
+
         game.refresh_enhancement_counts();
-        
+
         assert_eq!(game.steel_cards_in_deck, 4);
     }
 
     #[test]
     fn should_count_both_stone_and_steel_cards_independently() {
         let mut game = Game::new(Config::default());
-        
+
         let mixed_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Steel)),
@@ -4194,9 +4194,9 @@ mod stone_steel_tracking_tests {
             create_card_with_enhancement(Some(Enhancement::Bonus)), // Should be ignored
         ];
         game.add_cards_to_deck_for_testing(mixed_cards);
-        
+
         game.refresh_enhancement_counts();
-        
+
         assert_eq!(game.stone_cards_in_deck, 2);
         assert_eq!(game.steel_cards_in_deck, 3);
     }
@@ -4204,20 +4204,20 @@ mod stone_steel_tracking_tests {
     #[test]
     fn should_refresh_enhancement_counts_on_game_start() {
         let mut game = Game::new(Config::default());
-        
+
         let enhanced_cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Steel)),
         ];
         game.add_cards_to_deck_for_testing(enhanced_cards);
-        
+
         // Counts should be 0 before start (not yet calculated)
         assert_eq!(game.stone_cards_in_deck, 0);
         assert_eq!(game.steel_cards_in_deck, 0);
-        
+
         // Start the game (this should refresh counts)
         game.start();
-        
+
         // Counts should now be accurate
         assert_eq!(game.stone_cards_in_deck, 1);
         assert_eq!(game.steel_cards_in_deck, 1);
@@ -4226,9 +4226,9 @@ mod stone_steel_tracking_tests {
     #[test]
     fn should_handle_deck_size_calculation_correctly() {
         let mut game = Game::new(Config::default());
-        
+
         let initial_deck_size = game.deck_size_for_testing();
-        
+
         let cards = vec![
             create_card_with_enhancement(Some(Enhancement::Stone)),
             create_card_with_enhancement(Some(Enhancement::Stone)),
@@ -4238,7 +4238,7 @@ mod stone_steel_tracking_tests {
         ];
         game.add_cards_to_deck_for_testing(cards);
         game.refresh_enhancement_counts();
-        
+
         // Total deck size should be initial size + 5 added cards
         assert_eq!(game.deck_size_for_testing(), initial_deck_size + 5);
         // Enhancement counts should be accurate (only counting the added enhanced cards)