--- a/core/src/game/mod.rs
+++ b/core/src/game/mod.rs
@@ -456,6 +456,7 @@ impl Game {
                 stage: &self.stage,
                 hands_played: 0,
                 discards_used: 0,
+                hands_remaining: self.plays as f64,
                 jokers: &self.jokers,
                 hand: &temp_hand,
                 discarded: &self.discarded,
@@ -703,6 +704,7 @@ impl Game {
             stage: &self.stage,
             hands_played: 0,  // TODO: track this properly
             discards_used: 0, // TODO: track this properly
+            hands_remaining: self.plays as f64,
             jokers: &self.jokers,
             hand: &Hand::new(hand.hand.cards().to_vec()),
             discarded: &self.discarded,
@@ -873,6 +875,7 @@ impl Game {
                 stage: &self.stage,
                 hands_played: 0,  // TODO: track this properly
                 discards_used: 0, // TODO: track this properly
+                hands_remaining: self.plays as f64,
                 jokers: &self.jokers,
                 hand: &Hand::new(hand.hand.cards().to_vec()),
                 discarded: &self.discarded,
@@ -1209,6 +1212,7 @@ impl Game {
             stage: &self.stage,
             hands_played: (self.config.plays as f64 - self.plays) as u32,
             discards_used: (self.config.discards as f64 - self.discards) as u32,
+            hands_remaining: self.plays,
             jokers: &self.jokers,
             hand: &current_hand,
             discarded: &self.discarded,
@@ -2451,6 +2455,7 @@ impl Game {
             stage: &self.stage,
             hands_played: 0,  // TODO: track this properly
             discards_used: 0, // TODO: track this properly
+            hands_remaining: self.plays as f64,
             jokers: &self.jokers,
             hand: &crate::hand::Hand::new(vec![]),
             discarded: &self.discarded,