--- a/core/tests/effects_migration_test.rs
+++ b/core/tests/effects_migration_test.rs
@@ -212,6 +212,7 @@ fn test_migration_structured_effects_type_safety() {
         stage: &Stage::Blind(Blind::Small),
         hands_played: 0,
         discards_used: 0,
+        hands_remaining: 4.0, // Standard hands remaining for testing
         jokers: &[],
         hand: &temp_hand,
         discarded: &[],